Checkpoint before assistant change: Enable the analysis pipeline to run from the command line interface
Adds command line execution capability to TrendFinder/modules.py.
