+++ DataAnalyzer/analysis_pipeline.py
@@ -1,3 +1,5 @@
+"""I have some notes on it that I'd like you to make some changes with in mind if that's ok. First, I should have mentioned this upfront but this code that merges the categorization and analyzation together is actually going to be used as a smaller part of a bigger project itself. So its main use case will be in code, not terminal. So there's not much need to have command line tools I don't think, unless that's standard for python modules. Second, because of that I plan to define the lm object that dspy uses In an upper file that's responsible for calling all the parts together. This is to make sure that they all use the same lm object and I only have to change one instance if I decide to change the model or something. """
+
 #!/usr/bin/env python3
 """
 Unified Document Analysis Pipeline
