--- DataAnalyzer/pipeline_results_20250728_200927.csv
@@ -1,4 +1,0 @@
-pdf_document,re,hughes,adem_webpage_at_https,enclosure_person_name,permit_effective_date,permit_expiration_date,permittee_organization,organization mailing address city_id_number,organization_mailing_address_state,organization mailing address postal code_currency,inspection type_id_number,stormwater,inspection_reason,yes no weather conditions_person_name,site_id,site_name,activity_description,physical location city_id_number,physical_location_state,physical location postal code_currency,physical location county_id_id_number,physical_location_description,noi_entrance_latitude,fod_verified_entrance_lattitude,first name_id_number,last name_id_number,organization,position title_id_number,phone,email_address,mailing address city_id_number,mailing_address_state,mailing address postal code_currency,representatives_details,inspection_summary_inspection_begin_date,inspection_begin_time,inspection_end_date,inspection_end_time,summary_comments,field operations division office_id_number,inspector_person_name,phone_number_of_inspector,signature_of_inspector,reviewing supervisor_person_name,phone_number_of_reviewing_supervisor,signature_date,photo_taken,gps latitude_currency,gps_longitude,gps_azimuth,overall_photo_taken,maintenance_checklist_photo_taken,outfall_and_immediate_downstream_photo_taken,total_pages_processed,facility_inspection,csw_currency,cci_id_number,document_metadata
-TrendFinder/TrainingData/32037 ALR10C6BZ 097 07-10-2025 INSPR AEPACS NA.pdf,Facility Inspection – Minor Deficiencies Noted,Hughes Plumbing & Utility Contractors, Inc.,https://adem.alabama.gov/compliance-assistance,Felicia Smith,02/19/2025,03/31/2026,Hughes Plumbing & Utility Contractors, Inc.,Mobile,AL,36612,Construction Stormwater,CSW,CCI - Citizen Complaint/Tip,No, Felicia Smith,311J-P6F8-9TD,Big Creek Lake Dam Toe Berm and Drainage Improvements,Big Creek Dam erosion and drainage improvements,Wilmer,AL,36587,Mobile,Approximately 3 miles west of Snow on Tanner Williams Road (Big Creek Dam),,,Preston,Hughes,Hughes Plumbing & Utility Contractors, Inc.,President,2516109674,phughes@hpuconline.com,Mobile,AL,36612,No onsite representatives present,07/10/2025,12:10 PM,07/10/2025,2:55 PM,"Minor deficiencies noted, including improperly implemented/maintained BMPs (silt fencing), no structural BMPs around stockpiles, observed sedimentation, and onsite erosion.",Mobile,Felicia Smith,2512065404,Signed,Tim Connole,,07/14/2025,7/10/2025 1:56:36 PM,30.7121666666667,-88.3327233333333,9.14 degrees,7/10/2025 12:18:55 PM,7/10/2025 12:22:36 PM,7/10/2025 12:38:18 PM,10,Minor Deficiencies Noted,CSW,7M-004XU2C14,"CreationDate: D:20250714134428-05'00', ModDate: D:20250714134428-05'00', Producer: Aspose.PDF for .NET 24.11.0, Creator: Aspose Pty Ltd."
-TrendFinder/TrainingData/52781 ASN62022-CSW 097 07-10-2025 INSPR AEPACS Deer Crest.pdf,Facility Inspection – No Deficiencies Noted,,https://adem.alabama.gov/compliance-assistance,Felicia Smith,,,Deer Crest Property Owner's Association, Inc,Semmes,AL,36575,Construction Stormwater,CSW,CPI - Core Program,,62022-CSW,Deer Crest,Residential Subdivision Development,Mobile,AL,36608,Mobile,"Beginning southbound on I65 South, take Exit 0 onto I10 West toward Mississippi. Go 7.1 miles and take Exit 13 toward Theodore Dawes. Keep right at the fork and follow the signs for Dawes for 1.7 miles. Turn left onto Theodore Dawes Road West and drive for 1.1 miles, then turn right onto McDonald Road. Drive for 0.6 Miles, then turn left onto Three Notch Road. Drive for 0.5 Miles until you reach the traffic circle and take the second exit to stay on Three Notch Road. Drive on Three Notch Road for 1.5 Miles, then turn left onto Dawes Road. Follow Dawes Road for 3.7 miles, then turn right onto Hi Fields Road. Drive for 0.8 miles, then turn right onto Grand Bay Wilmer Road South. Stay on Grand Bay Wilmer Road South for 2.5 miles, then turn left onto South Grand Bay Wilmer Road. Drive for 1 mile, and then turn right into the Deer Crest Subdivision.,,30.6189883333333,Paul,Powers,Charter Property Owner's Association,President,2513420600,paul@powerland.biz,Semmes,AL,36575,No onsite representatives present,07/10/2025,2:15 PM,07/10/2025,2:43 PM,"According to NOI, this permit is for land, road, and drainage for residential subdvision. Road and drainage appeared to be complete. This expired permit may be eligible for closure.",Mobile,Felicia Smith,2512065404,Signed,Tim Connole,,07/14/2025,7/10/2025 2:15:50 PM,30.6189883333333,-88.359235,311.97 degrees,,,6,No Deficiencies Noted,CSW,,"CreationDate: D:20250714134426-05'00', ModDate: D:20250714134426-05'00', Author: , Producer: Aspose.PDF for .NET 24.11.0, Subject: , Title: , Creator: Aspose Pty Ltd."
-TrendFinder/TrainingData/61763 ALR10C6PZ 077 07-10-2025 INSPR AEPACS NA.pdf,Facility Inspection – Minor Deficiencies Noted,,https://adem.alabama.gov/compliance-assistance,Cal Johnson,05/19/2025,03/31/2026,WR Cove, LLC,FLORENCE,AL,356306403,Construction Stormwater - CSW,CSW,CPI - Core Program,Cloudy,Y3QD-81X9-3XJ,25-180 Natures Cove,16 home subdivision construction,FLORENCE,AL,356342113,Lauderdale,Alabama Department-Environmental 2715 Sandlin Rd SW, Decatur, AL 35603,,,Darren,Rhodes,WR Cove, LLC,Owner,2563357512,darren@calltrp.com,FLORENCE,AL,356306403,Mr. Clemmons,07/10/2025,9:30 AM,07/10/2025,10:20 AM,"No gravel at enter/exit pad. Silt fence compromised near lake edge. No permit posted nor rain gauge.",Decatur,Cal Johnson,2564322162,Signed,Shawn LaGrone,256-432-2076,07/11/2025,7/10/2025 2:56:00 PM,34.827858,87.582633,,7/10/2025 2:56:00 PM,7/10/2025 3:02:00 PM,,8,Minor Deficiencies Noted,CSW,,"CreationDate: D:20250711093448-05'00', ModDate: D:20250711093448-05'00', Author: , Producer: Aspose.PDF for .NET 24.11.0, Subject: , Title: , Creator: Aspose Pty Ltd."
